Support remove search ads
