+++ app/src/main/java/me/iacn/biliroaming/hook/BangumiSeasonHook.kt
@@ -26,7 +26,6 @@ import me.iacn.biliroaming.utils.*
 import org.json.JSONObject
 import java.io.InputStream
 import java.lang.reflect.Method
-import java.lang.reflect.Proxy
 import java.net.URL
 import java.net.URLDecoder
 import java.util.*
@@ -53,7 +52,7 @@ class BangumiSeasonHook(classLoader: ClassLoader) : BaseHook(classLoader) {
 
         data class Area(val area: String, val text: String, val type: String, val type_str: String)
 
-        private val AREA_TYPES =
+        val AREA_TYPES =
             mapOf(
                 931 to Area("cn", "陆(影)", "8", "movie"),
                 364364 to Area("hk", "港(影)", "8", "movie"),
@@ -99,6 +98,7 @@ class BangumiSeasonHook(classLoader: ClassLoader) : BaseHook(classLoader) {
             javaClass.getStaticObjectField("Companion")?.callMethod("serializer"),
             this
         ).toJSONObject()
+
         isGson -> gson?.callMethodAs<String>(instance.gsonToJson(), this)?.toJSONObject()
         else -> instance.fastJsonClass?.callStaticMethodAs<String>(
             "toJSONString",
@@ -113,6 +113,7 @@ class BangumiSeasonHook(classLoader: ClassLoader) : BaseHook(classLoader) {
             getStaticObjectField("Companion")?.callMethod("serializer"),
             json
         )
+
         isGson -> gson?.callMethod(instance.gsonFromJson(), json, this)
         else -> instance.fastJsonClass?.callStaticMethod(instance.fastJsonParse(), json, this)
     }
@@ -161,9 +162,6 @@ class BangumiSeasonHook(classLoader: ClassLoader) : BaseHook(classLoader) {
             mClassLoader
         )
     }
-
-    private val navClass by Weak { "com.bapis.bilibili.polymer.app.search.v1.Nav" from mClassLoader }
-    private val searchByTypeRespClass by Weak { "com.bapis.bilibili.polymer.app.search.v1.SearchByTypeResponse" from mClassLoader }
 
     @SuppressLint("SetTextI18n")
     override fun startHook() {
@@ -412,234 +410,6 @@ class BangumiSeasonHook(classLoader: ClassLoader) : BaseHook(classLoader) {
             }
             pageTypesClass.setStaticObjectField("\$VALUES", newPageArray)
         }
-        if (sPrefs.getBoolean("hidden", false) &&
-            (sPrefs.getBoolean("search_area_bangumi", false)
-                    || sPrefs.getBoolean("search_area_movie", false))
-        ) {
-            val mossResponseHandlerClass = instance.mossResponseHandlerClass ?: return
-            val searchMossClass =
-                "com.bapis.bilibili.polymer.app.search.v1.SearchMoss".from(mClassLoader) ?: return
-            searchMossClass.hookBeforeMethod(
-                "searchAll",
-                "com.bapis.bilibili.polymer.app.search.v1.SearchAllRequest",
-                mossResponseHandlerClass
-            ) { param ->
-                val handler = param.args[1]
-                param.args[1] = Proxy.newProxyInstance(
-                    handler.javaClass.classLoader,
-                    arrayOf(mossResponseHandlerClass)
-                ) { _, m, args ->
-                    if (m.name == "onNext") {
-                        addAreaTagsV2(args[0])
-                        m(handler, *args)
-                    } else if (args == null) {
-                        m(handler)
-                    } else {
-                        m(handler, *args)
-                    }
-                }
-            }
-            searchMossClass.hookBeforeMethod(
-                "searchByType",
-                "com.bapis.bilibili.polymer.app.search.v1.SearchByTypeRequest",
-                mossResponseHandlerClass
-            ) { param ->
-                val searchByTypeRespClass = searchByTypeRespClass ?: return@hookBeforeMethod
-                val key =
-                    param.args[0].callMethodOrNullAs<Int>("getType") ?: return@hookBeforeMethod
-                val areaType = AREA_TYPES[key] ?: return@hookBeforeMethod
-                val request = SearchByTypeRequest.parseFrom(
-                    param.args[0].callMethodAs<ByteArray>("toByteArray")
-                )
-                val type = areaType.type
-                val area = areaType.area
-                val handler = param.args[1]
-                MainScope().launch(Dispatchers.IO) {
-                    val result = retrieveAreaSearchV3(request, area, type)
-                    if (result != null) {
-                        val newRes = searchByTypeRespClass
-                            .callStaticMethod("parseFrom", result.toByteArray())
-                        handler.callMethod("onNext", newRes)
-                        handler.callMethod("onCompleted")
-                    } else {
-                        handler.callMethod("onError", null)
-                    }
-                }
-                param.result = null
-            }
-        }
-    }
-
-    private fun addAreaTagsV2(v: Any?) {
-        v ?: return
-        val navClass = navClass ?: return
-        val navList = v.callMethodAs<List<Any>>("getNavList")
-            .map { SearchNav.parseFrom(it.callMethodAs<ByteArray>("toByteArray")) }
-            .toMutableList()
-        val currentArea = runCatching {
-            XposedInit.country.get(5L, TimeUnit.SECONDS)
-        }.getOrNull()
-        for (area in AREA_TYPES) {
-            if (area.value.area == currentArea)
-                continue
-            if (!sPrefs.getString(area.value.area + "_server", null).isNullOrBlank() &&
-                sPrefs.getBoolean("search_area_" + area.value.type_str, false)
-            ) {
-                val nav = searchNav {
-                    name = area.value.text
-                    total = 0
-                    pages = 0
-                    type = area.key
-                }
-                navList.add(1, nav)
-            }
-        }
-        v.callMethod("clearNav")
-        val newNavList = navList.map {
-            navClass.callStaticMethod("parseFrom", it.toByteArray())
-        }
-        v.callMethod("addAllNav", newNavList)
-    }
-
-    private fun retrieveAreaSearchV3(
-        request: SearchByTypeRequest,
-        area: String,
-        type: String
-    ): SearchByTypeResponse? {
-        val pn = request.pagination.next.ifEmpty { "1" }
-        val ps = request.pagination.pageSize
-        val keyword = request.keyword
-        val query = mapOf(
-            "access_key" to (instance.accessKey ?: ""),
-            "fnval" to request.playerArgs.fnval,
-            "fnver" to request.playerArgs.fnver,
-            "qn" to request.playerArgs.qn,
-            "pn" to pn,
-            "ps" to ps,
-            "keyword" to keyword,
-        ).map { "${it.key}=${it.value}" }.joinToString("&")
-        val jsonContent = getAreaSearchBangumi(query, area, type)?.toJSONObject()
-            ?: return null
-        checkErrorToast(jsonContent, true)
-        val newData = jsonContent.optJSONObject("data") ?: return null
-
-        fun ReasonStyleKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            text = optString("text")
-            textColor = optString("text_color")
-            textColorNight = optString("text_color_night")
-            bgColor = optString("bg_color")
-            bgColorNight = optString("bg_color_night")
-            borderColor = optString("border_color")
-            borderColorNight = optString("border_color_night")
-            bgStyle = optInt("bg_style")
-        }
-
-        fun EpisodeKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            uri = optString("uri")
-            param = optString("param")
-            index = optString("index")
-            for (badge in optJSONArray("badges").orEmpty())
-                badges += reasonStyle { reconstructFrom(badge) }
-            position = optInt("position")
-        }
-
-        fun EpisodeNewKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            title = optString("title")
-            uri = optString("uri")
-            param = optString("param")
-            isNew = optInt("is_new")
-            for (badge in optJSONArray("badges").orEmpty())
-                badges += reasonStyle { reconstructFrom(badge) }
-            this@reconstructFrom.type = optInt("type")
-            position = optInt("position")
-            cover = optString("cover")
-            label = optString("label")
-        }
-
-        fun WatchButtonKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            title = optString("title")
-            link = optString("link")
-        }
-
-        fun CheckMoreKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            content = optString("content")
-            uri = optString("uri")
-        }
-
-        fun FollowButtonKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            icon = optString("icon")
-            optJSONObject("texts")?.let { o ->
-                o.keys().asSequence().associateWith { o.opt(it)?.toString() ?: "" }
-            }?.let { texts.putAll(it) }
-            statusReport = optString("status_report")
-        }
-
-        fun SearchBangumiCardKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            title = optString("title")
-            cover = optString("cover")
-            mediaType = optInt("media_type")
-            playState = optInt("play_state")
-            this@reconstructFrom.area = optString("area")
-            style = optString("style")
-            styles = optString("styles")
-            cv = optString("cv")
-            rating = optDouble("rating")
-            vote = optInt("vote")
-            target = optString("target")
-            staff = optString("staff")
-            prompt = optString("prompt")
-            ptime = optLong("ptime")
-            seasonTypeName = optString("season_type_name")
-            for (episode in optJSONArray("episodes").orEmpty())
-                episodes += episode { reconstructFrom(episode) }
-            isSelection = optInt("is_selection")
-            isAtten = optInt("is_atten")
-            label = optString("label")
-            seasonId = optLong("season_id")
-            outName = optString("out_name")
-            outIcon = optString("out_icon")
-            outUrl = optString("out_url")
-            for (badge in optJSONArray("badges").orEmpty())
-                badges += reasonStyle { reconstructFrom(badge) }
-            isOut = optInt("is_out")
-            for (episodeNew in optJSONArray("episodes_new").orEmpty())
-                episodesNew += episodeNew { reconstructFrom(episodeNew) }
-            optJSONObject("watch_button")?.let {
-                watchButton = watchButton { reconstructFrom(it) }
-            }
-            selectionStyle = optString("selection_style")
-            optJSONObject("check_more")?.let {
-                checkMore = checkMore { reconstructFrom(it) }
-            }
-            optJSONObject("follow_button")?.let {
-                followButton = followButton { reconstructFrom(it) }
-            }
-        }
-
-        fun SearchItemKt.Dsl.reconstructFrom(json: JSONObject) = json.run {
-            uri = optString("uri")
-            param = optString("param")
-            goto = optString("goto")
-            linkType = optString("link_type")
-            position = optInt("position")
-            trackId = optString("track_id")
-            bangumi = searchBangumiCard { reconstructFrom(json) }
-        }
-
-        val pages = newData.optInt("pages")
-        var page = pn.toIntOrNull() ?: 1
-        val response = searchByTypeResponse {
-            this.pages = pages
-            this.keyword = keyword
-            for (json in newData.optJSONArray("items").orEmpty()) {
-                if (json.optInt("Offset", -1) != -1)
-                    json.remove("follow_button")
-                items += searchItem { reconstructFrom(json) }
-            }
-            if (page < pages)
-                pagination = paginationReply { next = (++page).toString() }
-        }
-        return response
     }
 
     private fun fixSpace(mid: Long?): Any? {
