+++ app/src/main/res/xml/prefs_setting.xml
@@ -351,6 +351,12 @@
             android:key="block_up_rcmd_ads"
             android:summary="@string/block_up_rcmd_ads_summary"
             android:title="@string/block_up_rcmd_ads_title" />
+
+        <SwitchPreference
+            android:dependency="hidden"
+            android:key="search_remove_ads"
+            android:summary="@string/search_remove_ads_summary"
+            android:title="@string/search_remove_ads_title" />
     </PreferenceCategory>
 
     <PreferenceCategory
